chore: deprecate app v1
