+++ app/app_module_config.go
@@ -1,5 +1,3 @@
-//go:build !app_v1
-
 package app
 
 import (
